pyproject.toml
[project]
name = "mcp-servers"
version = "0.1.0"
description = "Collection of MCP servers for Oracle Cloud, Atlassian, and Code Repos"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Zhuo Li" }
]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "oci>=2.100.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
oracle-cloud-mcp = "mcp_servers.oracle_cloud.server:main"
atlassian-mcp = "mcp_servers.atlassian.server:main"
code-repos-mcp = "mcp_servers.code_repos.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_servers"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
